+++ PrebidMobile/PrebidMobile-core/src/main/java/org/prebid/mobile/api/multiadloader/MultiBannerLoader.kt
@@ -191,6 +191,7 @@ class MultiBannerLoader(
                     }
                 }
 
+                SdkLogUtil.info("GAM Ad requested", SdkAdStatus.REQUESTED, AdFormat.BANNER, gamAdUnitId, SdkType.GAM)
                 val request = AdManagerAdRequest.Builder().build()
                 loadAd(request)
             }